Update libpgf to official 6.09.24 including all my previous fixes !!! Thanks to libpgf team...
svn path=/trunk/extragear/graphics/digikam/; revision=981170

+++ libs/3rdparty/libpgf/Encoder.h
@@ -92,7 +92,7 @@ public:
 	/// It might throw an IOException.
 	/// @param band A subband
 	/// @param bandPosition A valid position in subband band
-	void WriteValue(CSubband* band, long bandPos) THROW_;
+	void WriteValue(CSubband* band, int bandPos) THROW_;
 
 #ifdef __PGFROISUPPORT__
 	/////////////////////////////////////////////////////////////////////
